refactor: 수정된 파일 경로 반영
